refactor : remove Service Interface
